requirements.txt
# Core
torch
transformers
accelerate
bitsandbytes
scipy

# Vision & Data
opencv-python-headless
pandas
scikit-learn
tqdm
Pillow
requests

# Hugging Face
huggingface_hub